chore: gate commits on ruff format, ruff check and ty via pre-commit
Three local hooks run through `uv run` so they use the same locked tool
versions as the verify steps and cannot drift from them; commits that touch
no Python skip all three.

ty joins the dev group (0.0.80). fontTools' otTables and uharfbuzz assemble
members at runtime and ship no stubs, so unresolved-attribute is downgraded
per-file in kern_axis/specimen/test_shaping, GlyphsApp is allowed only where
it exists (build.py, macro_bootstrap.py), and error-on-warning = false keeps
the ~51 stub-gap warnings visible but non-gating.

Real diagnostics fixed: _ink_y_span return annotation (floats, not ints),
getBestCmap() None-narrowing asserts in demo.py/specimen.py and the tests,
and a guarded regex Match in test_demo.

+++ src/monolith/kerning.py
@@ -91,7 +91,7 @@ def _ink_extent(shapes: tuple, y: float) -> tuple[float, float] | None:
     return min(a for a, _ in iv), max(b for _, b in iv)
 
 
-def _ink_y_span(name: str) -> tuple[int, int] | None:
+def _ink_y_span(name: str) -> tuple[float, float] | None:
     """(lowest y, highest y) of a glyph's ink, or None for the empty glyph."""
     ys = [py for s in DES[name].shapes for (_, py) in _shape_points(s)]
     return (min(ys), max(ys)) if ys else None
